+++ Frameworks/libsidplay/sidplay-residfp-code/libsidplayfp/src/c64/CIA/tod.cpp
@@ -134,7 +134,7 @@ void Tod::event()
     cycles += period * (cra & 0x80 ? 5 : 6);
 
     // Fixed precision 25.7
-    eventScheduler.schedule(*this, cycles >> 7);
+    eventScheduler.schedule(*this, (unsigned int)(cycles >> 7));
     cycles &= 0x7F; // Just keep the decimal part
 
     if (!isStopped)
