+++ Frameworks/libsidplay/sidplay-residfp-code/libsidplayfp/src/c64/VIC_II/mos656x.cpp
@@ -274,7 +274,7 @@ void MOS656X::event()
     else
         delay = 1;
 
-    eventScheduler.schedule(*this, delay - eventScheduler.phase(), EVENT_CLOCK_PHI1);
+    eventScheduler.schedule(*this, (unsigned int)(delay - eventScheduler.phase()), EVENT_CLOCK_PHI1);
 }
 
 event_clock_t MOS656X::clockPAL()
